% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_examples.R
\name{get_examples}
\alias{get_examples}
\title{Get example CSV files}
\usage{
get_examples()
}
\description{
This function gives access to the CSV files given as demonstrations when
using the \code{\link{learn}} function.
}
\examples{
get_examples()

}
