% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnDir.R
\name{learnDir}
\alias{learnDir}
\title{Open the datasets directory}
\usage{
learnDir(polyglotDirectory = paste0("", system.file("extdata/", package =
  "polyglot"), ""))
}
\arguments{
\item{polyglotDirectory}{A string}
}
\description{
This function will open the directory which contains the dataset files.
}
\details{
This function opens the directory which contain all the dataset files which are
loaded in polyglot.
}
\examples{
if(interactive()){ 
learnDir() 
}

}
\references{
A blog post describing the package with more development: \url{https://lgnbhl.github.io/learner}
}
