% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionExit.R
\name{sessionExit}
\alias{sessionExit}
\title{Exit the interactive environment}
\usage{
sessionExit(assign.env = parent.frame(1))
}
\arguments{
\item{assign.env}{An environment}
}
\description{
This function will quit the interactive learning environment.
}
\details{
This function calculates the learning session time and print it. Then, it
cleans the Global Environment if needed and displays a farewell message.
}
