% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_dir.R
\name{learn_dir}
\alias{learn_dir}
\title{Open the datasets directory}
\usage{
learn_dir(polyglotDirectory = pins::board_local_storage())
}
\arguments{
\item{polyglotDirectory}{A string}
}
\description{
This function will open the directory which contains the local datasets files 
used in polygot thanks to the pins package.
}
\details{
This function opens the directory which contain all the dataset files which are
loaded into the interactive learning environment of the polyglot package.
}
\examples{
if(interactive()){
learn_dir()
}

}
